requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.44.1
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
Eel==0.17.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.7
lxml==5.2.2
packaging==24.1
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
python-docx==1.1.2
pywin32-ctypes==0.2.2
requests==2.32.3
setuptools==70.3.0
typing_extensions==4.12.2
urllib3==2.2.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.4.post2
